console.log() weggehaald, code opgekuist, code duplicatie vermeden.
